requirements.txt
requests
sentence-transformers
faiss-cpu
python-dotenv